Reformated engines into folders w/ their own files. Fixed small bug regarding finding context.
